+++ src/cli/default-error-handler.js
@@ -23,6 +23,7 @@ import { ScopeNotFound, ResolutionException, ComponentNotFound, DependencyNotFou
 import { ProtocolNotSupported, RemoteScopeNotFound } from '../scope/network/exceptions';
 import InvalidBitJson from '../consumer/bit-json/exceptions/invalid-bit-json';
 import ExportWithoutThis from '../api/consumer/lib/exceptions/export-without-this';
+import invalidIdOnCommit from '../api/consumer/lib/exceptions/invalid-id-on-commit';
 import MiscSourceNotFound from '../consumer/component/exceptions/misc-source-not-found';
 
 const errorsMap: [[Error, (err: Error) => string]] = [
@@ -50,7 +51,9 @@ const errorsMap: [[Error, (err: Error) => string]] = [
   [ InvalidBitJson, err => `error: ${chalk.bold(err.path)} is not a valid JSON file.`],
   [ ExportWithoutThis, err => `Error: Missing local scope annotation when exporting ${err.id}. Please run 'bit export ${LOCAL_SCOPE_NOTATION}/${err.id} ${err.remote}'`],
   [ ResolutionException, e => e.message],
-  [ DriverNotFound, err => `fatal: a client-driver ${chalk.bold(err.driver)} is missing for the language ${chalk.bold(err.lang)} set in your bit.json file.`]
+  [ DriverNotFound, err => `fatal: a client-driver ${chalk.bold(err.driver)} is missing for the language ${chalk.bold(err.lang)} set in your bit.json file.`],
+  [ invalidIdOnCommit, err => `error - Unable to commit. ${chalk.bold(err.id)} not found.
+Run \`bit status\` command to list all components available for commit.`]
 ];
 
 export default (err: Error): ?string => {