+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- send correct error message when there is invalid inline id (wip)
-- add bind command (which calls the driver bind command)
+- do not write bit.json's `misc` and `lang` properties if the default value is presented.
+- send correct error message when there is invalid inline id (wip).
+- add bind command (which calls the driver bind command).
 
 ## [0.5.10] - 2017-05-11 rc