+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - add onModify hook
+- show error-message for 'bit build' when no compiler is specified 
+- write dependencies on modify.
 - do not write bit.json's `misc` and `lang` properties if the default value is presented.
 - send correct error message when there is invalid inline id (wip).
 - add bind command (which calls the driver bind command).